Comment out a non-public Kahuna method
This method's sole purpose is to tell Kahuna the company that is wrapping their SDK and the version of the wrapper. Since this is not a public method it raises a compiler warning. We will keep it commented out until the respective method is made public in the Kahuna SDK.

There is no side effect in the integration functionality